refactor: move course object from core to root
