+++ validator/core/__init__.py
@@ -1,5 +1,5 @@
 from tokenize import Single
 from .solver import GraduationRequirementsSolver
-from .utils import Course, SingleAssignment
+from .utils import SingleAssignment
 
-__all__ = ["GraduationRequirementsSolver", "Course", "SingleAssignment"]
+__all__ = ["GraduationRequirementsSolver", "SingleAssignment"]
