+++ validator/core/solver.py
@@ -16,6 +16,8 @@ from functools import reduce
 import os
 from dotenv import load_dotenv
 
+from course import Course
+
 load_dotenv()
 
 
@@ -256,16 +258,7 @@ class GraduationRequirementsSolver:
         # Ensure requirements are valid
         self.validate()
 
-    def solve(self, courses: list[str], bypasses: list[SingleAssignment]):
-        # Convert list of str into list of courses
-        newCourses = []
-        for course in courses:
-            sub_prefix, course_num = course.split(" ")
-            newCourses.append(
-                Course(course, int(course_num[0]), int(course_num[1]), sub_prefix)
-            )
-        courses = newCourses
-
+    def solve(self, courses: list[Course], bypasses: list[SingleAssignment]):
         # Pre-process bypasses into an assignment, and validate them
         bypass_assignments = AssignmentStore()
         courses_dict: dict[str, Course] = {course.name: course for course in courses}
