+++ validator/major/requirements/base.py
@@ -16,7 +16,10 @@ class AbstractRequirement(ABC):
     """
 
     @abstractmethod
-    def attempt_fulfill(self, course: str) -> bool:
+    def attempt_fulfill(
+        self,
+        course: str,
+    ) -> bool:
         pass
 
     @abstractmethod
